+++ apps/web/src/edge/tokenPrice.ts
@@ -8,7 +8,7 @@ import { CAKE, getTokenByAddress, STABLE_COIN } from '@pancakeswap/tokens'
 import { chainlinkOracleABI } from 'config/abi/chainlinkOracle'
 import { getMulticallGasLimit } from 'quoter/hook/useMulticallGasLimit'
 import { edgeQueries } from 'quoter/utils/edgePoolQueries'
-import { computeTradePriceBreakdown, warningSeverity } from 'utils/exchange'
+import { computeTradePriceBreakdown, warningSeverity } from 'utils/compuateTradePriceBreakdown'
 import { getViemClients } from 'utils/viem.server'
 import { Address } from 'viem/accounts'
 import { formatUnits } from 'viem/utils'
