+++ apps/web/src/utils/compuateTradePriceBreakdown.ts
@@ -1,0 +1,54 @@
+import { Currency, CurrencyAmount, Fraction, Percent, Trade, TradeType } from '@pancakeswap/sdk'
+import {
+  ALLOWED_PRICE_IMPACT_HIGH,
+  ALLOWED_PRICE_IMPACT_LOW,
+  ALLOWED_PRICE_IMPACT_MEDIUM,
+  BLOCKED_PRICE_IMPACT_NON_EXPERT,
+  INPUT_FRACTION_AFTER_FEE,
+  ONE_HUNDRED_PERCENT,
+} from 'config/constants/exchange'
+
+// computes price breakdown for the trade
+export function computeTradePriceBreakdown(trade?: Trade<Currency, Currency, TradeType> | null): {
+  priceImpactWithoutFee: Percent | undefined
+  realizedLPFee: CurrencyAmount<Currency> | undefined | null
+} {
+  // for each hop in our trade, take away the x*y=k price impact from 0.3% fees
+  // e.g. for 3 tokens/2 hops: 1 - ((1 - .03) * (1-.03))
+  const realizedLPFee = !trade
+    ? undefined
+    : ONE_HUNDRED_PERCENT.subtract(
+        trade.route.pairs.reduce<Fraction>(
+          (currentFee: Fraction): Fraction => currentFee.multiply(INPUT_FRACTION_AFTER_FEE),
+          ONE_HUNDRED_PERCENT,
+        ),
+      )
+
+  // remove lp fees from price impact
+  const priceImpactWithoutFeeFraction = trade && realizedLPFee ? trade?.priceImpact.subtract(realizedLPFee) : undefined
+
+  // the x*y=k impact
+  const priceImpactWithoutFeePercent = priceImpactWithoutFeeFraction
+    ? new Percent(priceImpactWithoutFeeFraction?.numerator, priceImpactWithoutFeeFraction?.denominator)
+    : undefined
+
+  // the amount of the input that accrues to LPs
+  const realizedLPFeeAmount =
+    realizedLPFee &&
+    trade &&
+    CurrencyAmount.fromRawAmount(
+      trade.inputAmount.currency,
+      realizedLPFee.multiply(trade.inputAmount.quotient).quotient,
+    )
+
+  return { priceImpactWithoutFee: priceImpactWithoutFeePercent, realizedLPFee: realizedLPFeeAmount }
+}
+
+export function warningSeverity(priceImpact: Percent | undefined | null): 0 | 1 | 2 | 3 | 4 {
+  if (!priceImpact) return 0
+  if (!priceImpact?.lessThan(BLOCKED_PRICE_IMPACT_NON_EXPERT)) return 4
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_HIGH)) return 3
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_MEDIUM)) return 2
+  if (!priceImpact?.lessThan(ALLOWED_PRICE_IMPACT_LOW)) return 1
+  return 0
+}
